Move files into the right place
